+++ src/main/java/homework_2/GarageClass.java
