+++ src/main/java/homework_2/Owner.java
