+++ src/main/java/homework_2/utility/ComparePowerUtility.java
